chapter selection addition procedure
I was capable of making the chapter selection button/hover thingy work in both the phone and the computer. The control pad will change colour and only display the chapters.

However the chapters are not yet displayed, because I am tired and it's 1:30 I need to go to sleep. This is great progress however, it worked better than expected.